requirements.txt
nicegui
pandas
requests
numpy
google-generativeai
aiohttp